fix(broker): pr #845 round-6 triangulation — directory split cleanup
Round-6 fix for the flattened combined PR, addressing findings from a
3-lens codex triangulation (architecture, api, sre) of the post-split
directory structure. The directory split landed in 5 prior commits;
this round tightens module boundaries that the structure review
flagged.

Changes
- Drop internal-type re-exports from `replay-check/index.ts`. The
  subsystem barrel had been re-exporting `ReplayedBudget` and
  `BudgetCandidateIndexes` for "type inference convenience", but
  no consumer needs them and the package barrel deliberately scopes
  the public surface to `ReplayCheckReport` / `ReplayDiscrepancy` /
  `runReplayCheck`. Two competing "public" stories collapse to one.
  Convergent 3-of-3 finding (architecture MEDIUM, api LOW, sre LOW).
- Move `__createBudgetCandidateIndexesForTesting` definition into
  `testing.ts`. Production module `budget-candidate-index.ts` no
  longer exports an `__ForTesting` alias; the test seam owns its
  own constructor surface. Convergent 2-of-3 (architecture, sre).
- Split `internal.ts` from a 126-LOC junk drawer into a 36-LOC
  strict-shared-only module. Items used by exactly one consumer
  moved next to that consumer:
    - `BATCH_SIZE`, `CostEventBatchRow`, `AgentDayDbRow`, `TaskDbRow`,
      `HighestLsnRow`, `agentDayKey`, `eventTypeToKind` → orchestrator
      (`index.ts`)
    - `arraysEqual`, `parseStoredThresholds` → `budget-comparator.ts`
    - `splitAgentDayKey` → `aggregate-comparators.ts`
  Remaining shared items in `internal.ts`: `BudgetDbRow`,
  `ThresholdCrossingDbRow`, `crossingKey` — each imported by 2+
  sibling modules. New strict membership rule documented in the
  module header. Architecture lens LOW finding.
- Update pre-split comments that still pointed contributors at
  `replay-check.ts` (the old monolith). `computeExpectedCrossings`
  is now in `threshold-oracle.ts`; the test seam is `testing.ts`.
  Convergent 3-of-3 finding across all three lenses.
- Strip stale "extracted for file-size limit" comments — the
  decomposition is its own justification now.

Dispositions
- 3-of-3 (stale comments) — FIXED
- 3-of-3 (internal-type re-exports leak) — FIXED
- 2-of-3 (test-only factory in production module) — FIXED
- architecture (internal.ts junk-drawer risk) — FIXED via strict
  membership rule
- api MEDIUM (`budget_state_mismatch` per-field JSON shape is
  implicit in comparator code) — DEFERRED. Splitting the variant
  into typed per-field shapes is a meaningful wire-shape change
  that warrants its own scope.
- api MEDIUM (no documented upgrade policy for new
  `ReplayDiscrepancy` kinds) — DEFERRED to a follow-up doc PR.
