+++ packages/broker/src/cost-ledger/replay-check/aggregate-comparators.ts
@@ -7,7 +7,20 @@
 // the discrepancy wire shape is a decimal string, so the diagnostic
 // preserves exact bytes regardless of magnitude.
 import type { ReplayDiscrepancy } from "./discrepancy.ts";
-import { splitAgentDayKey } from "./internal.ts";
+
+// Inverse of the orchestrator's `agentDayKey` encoding. Only the
+// aggregate comparators round-trip from key back to its components,
+// so the helper lives here next to its single call site.
+function splitAgentDayKey(key: string): {
+  readonly agentSlug: string;
+  readonly dayUtc: string;
+} {
+  const idx = key.indexOf("|");
+  if (idx === -1) {
+    throw new Error(`replay-check: malformed agent-day key ${key}`);
+  }
+  return { agentSlug: key.slice(0, idx), dayUtc: key.slice(idx + 1) };
+}
 
 export function compareAgentDays(
   replayed: ReadonlyMap<string, bigint>,
